refactor(table): extract manifest assembly, leaving transaction.rs a facade
build_manifest and the helpers that feed it move to manifest_build, which
completes the split: transaction.rs is now module declarations, re-exports, and
a map of where each concern lives.

The re-export list is the same set of names the module exported before, so
nothing outside lance-table sees a change. Items used across submodules are
pub(super) rather than pub(crate), since the submodules themselves are private.
